Inspection rework
