+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Current
 - [Detect unset userPrincipal in Preface log block](https://github.com/yahoo/fili/pull/154)
     * Logs a warning if no userPrincipal is set on the request (ie. we don't know who the user is), and sets the
       `user` field in the `Preface` log block to `NO_USER_PRINCIPAL`.
+- [Support timeouts for lucene search provider](https://github.com/yahoo/fili/pull/183)
 
 ### Changed:
