+++ fili-core/src/main/resources/moduleConfig.properties
@@ -63,6 +63,9 @@ bard__druid_cache_v2_enabled = true
 # Lucene index files path
 bard__lucene_index_path = [SET ME IN APPLICATION CONFIG]
 
+# Lucene search timeout in milliseconds
+bard__lucene_search_timeout_ms = 600000
+
 # setting for maximum allowed results without any filters - used for /dim/values endpoint
 bard__max_results_without_filters = 10000
